+++ src/components/organisms/teskDialog/data/constants.ts
@@ -1,4 +1,4 @@
-import { AddTask } from '@/types/task/task.type';
+import { AddTask, Category } from '@/types/task/task.type';
 import { format } from 'date-fns';
 
 export const defaultAddTask: AddTask = {
@@ -14,3 +14,31 @@ export const defaultAddTask: AddTask = {
   },
   priority: 5,
 };
+
+// TODO:TESTCODE
+export const categories: Category[] = [
+  {
+    id: 1,
+    name: 'work',
+    color: 'bg-red-200',
+    icon: '🌈',
+  },
+  {
+    id: 2,
+    name: 'java study',
+    color: 'bg-amber-950',
+    icon: '🔥',
+  },
+  {
+    id: 3,
+    name: 'travel',
+    color: 'bg-yellow-200',
+    icon: '⚡',
+  },
+  {
+    id: 4,
+    name: 'work',
+    color: 'bg-blue-400',
+    icon: '🎶',
+  },
+];
